Related work items: #20113

+++ dotnet/corecomponents/net.atos.daf.ct2.fms/entity/VehiclePositionResponse.cs
@@ -13,11 +13,11 @@ namespace net.atos.daf.ct2.fms.entity
     {
         public string VIN { get; set; }
         public int Altitude { get; set; }
-        public double Heading { get; set; }
+        public int Heading { get; set; }
         public double Latitude { get; set; }
         public double Longitude { get; set; }
         public long GPSTimestamp { get; set; }
-        public double Speed { get; set; }
+        public decimal Speed { get; set; }
     }
 }
